GitHub users

Info on GitHub users, retrieved from the GitHub API.

gh_users

Format

A unnamed list with 6 components, each representing a GitHub user. Each user's component is a list of length 30, containing information such as username, GitHub id, and join date.

Source

https://developer.github.com/v3/users/#get-a-single-user

See also

Examples

gh_users
#> [[1]]
#> [[1]]$login
#> [1] "gaborcsardi"
#> 
#> [[1]]$id
#> [1] 660288
#> 
#> [[1]]$avatar_url
#> [1] "https://avatars.githubusercontent.com/u/660288?v=3"
#> 
#> [[1]]$gravatar_id
#> [1] ""
#> 
#> [[1]]$url
#> [1] "https://api.github.com/users/gaborcsardi"
#> 
#> [[1]]$html_url
#> [1] "https://github.com/gaborcsardi"
#> 
#> [[1]]$followers_url
#> [1] "https://api.github.com/users/gaborcsardi/followers"
#> 
#> [[1]]$following_url
#> [1] "https://api.github.com/users/gaborcsardi/following{/other_user}"
#> 
#> [[1]]$gists_url
#> [1] "https://api.github.com/users/gaborcsardi/gists{/gist_id}"
#> 
#> [[1]]$starred_url
#> [1] "https://api.github.com/users/gaborcsardi/starred{/owner}{/repo}"
#> 
#> [[1]]$subscriptions_url
#> [1] "https://api.github.com/users/gaborcsardi/subscriptions"
#> 
#> [[1]]$organizations_url
#> [1] "https://api.github.com/users/gaborcsardi/orgs"
#> 
#> [[1]]$repos_url
#> [1] "https://api.github.com/users/gaborcsardi/repos"
#> 
#> [[1]]$events_url
#> [1] "https://api.github.com/users/gaborcsardi/events{/privacy}"
#> 
#> [[1]]$received_events_url
#> [1] "https://api.github.com/users/gaborcsardi/received_events"
#> 
#> [[1]]$type
#> [1] "User"
#> 
#> [[1]]$site_admin
#> [1] FALSE
#> 
#> [[1]]$name
#> [1] "Gábor Csárdi"
#> 
#> [[1]]$company
#> [1] "Mango Solutions, @MangoTheCat "
#> 
#> [[1]]$blog
#> [1] "http://gaborcsardi.org"
#> 
#> [[1]]$location
#> [1] "Chippenham, UK"
#> 
#> [[1]]$email
#> [1] "csardi.gabor@gmail.com"
#> 
#> [[1]]$hireable
#> NULL
#> 
#> [[1]]$bio
#> NULL
#> 
#> [[1]]$public_repos
#> [1] 52
#> 
#> [[1]]$public_gists
#> [1] 6
#> 
#> [[1]]$followers
#> [1] 303
#> 
#> [[1]]$following
#> [1] 22
#> 
#> [[1]]$created_at
#> [1] "2011-03-09T17:29:25Z"
#> 
#> [[1]]$updated_at
#> [1] "2016-10-11T11:05:06Z"
#> 
#> 
#> [[2]]
#> [[2]]$login
#> [1] "jennybc"
#> 
#> [[2]]$id
#> [1] 599454
#> 
#> [[2]]$avatar_url
#> [1] "https://avatars.githubusercontent.com/u/599454?v=3"
#> 
#> [[2]]$gravatar_id
#> [1] ""
#> 
#> [[2]]$url
#> [1] "https://api.github.com/users/jennybc"
#> 
#> [[2]]$html_url
#> [1] "https://github.com/jennybc"
#> 
#> [[2]]$followers_url
#> [1] "https://api.github.com/users/jennybc/followers"
#> 
#> [[2]]$following_url
#> [1] "https://api.github.com/users/jennybc/following{/other_user}"
#> 
#> [[2]]$gists_url
#> [1] "https://api.github.com/users/jennybc/gists{/gist_id}"
#> 
#> [[2]]$starred_url
#> [1] "https://api.github.com/users/jennybc/starred{/owner}{/repo}"
#> 
#> [[2]]$subscriptions_url
#> [1] "https://api.github.com/users/jennybc/subscriptions"
#> 
#> [[2]]$organizations_url
#> [1] "https://api.github.com/users/jennybc/orgs"
#> 
#> [[2]]$repos_url
#> [1] "https://api.github.com/users/jennybc/repos"
#> 
#> [[2]]$events_url
#> [1] "https://api.github.com/users/jennybc/events{/privacy}"
#> 
#> [[2]]$received_events_url
#> [1] "https://api.github.com/users/jennybc/received_events"
#> 
#> [[2]]$type
#> [1] "User"
#> 
#> [[2]]$site_admin
#> [1] FALSE
#> 
#> [[2]]$name
#> [1] "Jennifer (Jenny) Bryan"
#> 
#> [[2]]$company
#> [1] "University of British Columbia"
#> 
#> [[2]]$blog
#> [1] "https://twitter.com/JennyBryan"
#> 
#> [[2]]$location
#> [1] "Vancouver, BC, Canada"
#> 
#> [[2]]$email
#> NULL
#> 
#> [[2]]$hireable
#> NULL
#> 
#> [[2]]$bio
#> [1] "prof at UBC, humane #rstats, statistics, teach @STAT545-UBC, leadership of @rOpenSci, @rsheets, academic director @ubc-mds"
#> 
#> [[2]]$public_repos
#> [1] 168
#> 
#> [[2]]$public_gists
#> [1] 54
#> 
#> [[2]]$followers
#> [1] 780
#> 
#> [[2]]$following
#> [1] 34
#> 
#> [[2]]$created_at
#> [1] "2011-02-03T22:37:41Z"
#> 
#> [[2]]$updated_at
#> [1] "2016-10-24T07:20:26Z"
#> 
#> 
#> [[3]]
#> [[3]]$login
#> [1] "jtleek"
#> 
#> [[3]]$id
#> [1] 1571674
#> 
#> [[3]]$avatar_url
#> [1] "https://avatars.githubusercontent.com/u/1571674?v=3"
#> 
#> [[3]]$gravatar_id
#> [1] ""
#> 
#> [[3]]$url
#> [1] "https://api.github.com/users/jtleek"
#> 
#> [[3]]$html_url
#> [1] "https://github.com/jtleek"
#> 
#> [[3]]$followers_url
#> [1] "https://api.github.com/users/jtleek/followers"
#> 
#> [[3]]$following_url
#> [1] "https://api.github.com/users/jtleek/following{/other_user}"
#> 
#> [[3]]$gists_url
#> [1] "https://api.github.com/users/jtleek/gists{/gist_id}"
#> 
#> [[3]]$starred_url
#> [1] "https://api.github.com/users/jtleek/starred{/owner}{/repo}"
#> 
#> [[3]]$subscriptions_url
#> [1] "https://api.github.com/users/jtleek/subscriptions"
#> 
#> [[3]]$organizations_url
#> [1] "https://api.github.com/users/jtleek/orgs"
#> 
#> [[3]]$repos_url
#> [1] "https://api.github.com/users/jtleek/repos"
#> 
#> [[3]]$events_url
#> [1] "https://api.github.com/users/jtleek/events{/privacy}"
#> 
#> [[3]]$received_events_url
#> [1] "https://api.github.com/users/jtleek/received_events"
#> 
#> [[3]]$type
#> [1] "User"
#> 
#> [[3]]$site_admin
#> [1] FALSE
#> 
#> [[3]]$name
#> [1] "Jeff L."
#> 
#> [[3]]$company
#> NULL
#> 
#> [[3]]$blog
#> [1] "http://biostat.jhsph.edu/~jleek/"
#> 
#> [[3]]$location
#> [1] "Baltimore,MD"
#> 
#> [[3]]$email
#> NULL
#> 
#> [[3]]$hireable
#> NULL
#> 
#> [[3]]$bio
#> NULL
#> 
#> [[3]]$public_repos
#> [1] 67
#> 
#> [[3]]$public_gists
#> [1] 12
#> 
#> [[3]]$followers
#> [1] 3958
#> 
#> [[3]]$following
#> [1] 6
#> 
#> [[3]]$created_at
#> [1] "2012-03-24T18:16:43Z"
#> 
#> [[3]]$updated_at
#> [1] "2016-10-05T13:03:04Z"
#> 
#> 
#> [[4]]
#> [[4]]$login
#> [1] "juliasilge"
#> 
#> [[4]]$id
#> [1] 12505835
#> 
#> [[4]]$avatar_url
#> [1] "https://avatars.githubusercontent.com/u/12505835?v=3"
#> 
#> [[4]]$gravatar_id
#> [1] ""
#> 
#> [[4]]$url
#> [1] "https://api.github.com/users/juliasilge"
#> 
#> [[4]]$html_url
#> [1] "https://github.com/juliasilge"
#> 
#> [[4]]$followers_url
#> [1] "https://api.github.com/users/juliasilge/followers"
#> 
#> [[4]]$following_url
#> [1] "https://api.github.com/users/juliasilge/following{/other_user}"
#> 
#> [[4]]$gists_url
#> [1] "https://api.github.com/users/juliasilge/gists{/gist_id}"
#> 
#> [[4]]$starred_url
#> [1] "https://api.github.com/users/juliasilge/starred{/owner}{/repo}"
#> 
#> [[4]]$subscriptions_url
#> [1] "https://api.github.com/users/juliasilge/subscriptions"
#> 
#> [[4]]$organizations_url
#> [1] "https://api.github.com/users/juliasilge/orgs"
#> 
#> [[4]]$repos_url
#> [1] "https://api.github.com/users/juliasilge/repos"
#> 
#> [[4]]$events_url
#> [1] "https://api.github.com/users/juliasilge/events{/privacy}"
#> 
#> [[4]]$received_events_url
#> [1] "https://api.github.com/users/juliasilge/received_events"
#> 
#> [[4]]$type
#> [1] "User"
#> 
#> [[4]]$site_admin
#> [1] FALSE
#> 
#> [[4]]$name
#> [1] "Julia Silge"
#> 
#> [[4]]$company
#> NULL
#> 
#> [[4]]$blog
#> [1] "juliasilge.com"
#> 
#> [[4]]$location
#> [1] "Salt Lake City, UT"
#> 
#> [[4]]$email
#> NULL
#> 
#> [[4]]$hireable
#> NULL
#> 
#> [[4]]$bio
#> [1] "Data science and visualization, #rstats, parenthood, reading, food/wine/coffee, #NASADatanauts."
#> 
#> [[4]]$public_repos
#> [1] 26
#> 
#> [[4]]$public_gists
#> [1] 4
#> 
#> [[4]]$followers
#> [1] 115
#> 
#> [[4]]$following
#> [1] 10
#> 
#> [[4]]$created_at
#> [1] "2015-05-19T02:51:23Z"
#> 
#> [[4]]$updated_at
#> [1] "2016-10-20T10:41:43Z"
#> 
#> 
#> [[5]]
#> [[5]]$login
#> [1] "leeper"
#> 
#> [[5]]$id
#> [1] 3505428
#> 
#> [[5]]$avatar_url
#> [1] "https://avatars.githubusercontent.com/u/3505428?v=3"
#> 
#> [[5]]$gravatar_id
#> [1] ""
#> 
#> [[5]]$url
#> [1] "https://api.github.com/users/leeper"
#> 
#> [[5]]$html_url
#> [1] "https://github.com/leeper"
#> 
#> [[5]]$followers_url
#> [1] "https://api.github.com/users/leeper/followers"
#> 
#> [[5]]$following_url
#> [1] "https://api.github.com/users/leeper/following{/other_user}"
#> 
#> [[5]]$gists_url
#> [1] "https://api.github.com/users/leeper/gists{/gist_id}"
#> 
#> [[5]]$starred_url
#> [1] "https://api.github.com/users/leeper/starred{/owner}{/repo}"
#> 
#> [[5]]$subscriptions_url
#> [1] "https://api.github.com/users/leeper/subscriptions"
#> 
#> [[5]]$organizations_url
#> [1] "https://api.github.com/users/leeper/orgs"
#> 
#> [[5]]$repos_url
#> [1] "https://api.github.com/users/leeper/repos"
#> 
#> [[5]]$events_url
#> [1] "https://api.github.com/users/leeper/events{/privacy}"
#> 
#> [[5]]$received_events_url
#> [1] "https://api.github.com/users/leeper/received_events"
#> 
#> [[5]]$type
#> [1] "User"
#> 
#> [[5]]$site_admin
#> [1] FALSE
#> 
#> [[5]]$name
#> [1] "Thomas J. Leeper"
#> 
#> [[5]]$company
#> [1] "London School of Economics"
#> 
#> [[5]]$blog
#> [1] "http://www.thomasleeper.com"
#> 
#> [[5]]$location
#> [1] "London, United Kingdom"
#> 
#> [[5]]$email
#> NULL
#> 
#> [[5]]$hireable
#> [1] TRUE
#> 
#> [[5]]$bio
#> [1] "Political scientist and R hacker. Interested in open science, public opinion research, surveys, experiments, crowdsourcing, and cloud computing."
#> 
#> [[5]]$public_repos
#> [1] 99
#> 
#> [[5]]$public_gists
#> [1] 46
#> 
#> [[5]]$followers
#> [1] 213
#> 
#> [[5]]$following
#> [1] 230
#> 
#> [[5]]$created_at
#> [1] "2013-02-07T21:07:00Z"
#> 
#> [[5]]$updated_at
#> [1] "2016-10-20T10:36:25Z"
#> 
#> 
#> [[6]]
#> [[6]]$login
#> [1] "masalmon"
#> 
#> [[6]]$id
#> [1] 8360597
#> 
#> [[6]]$avatar_url
#> [1] "https://avatars.githubusercontent.com/u/8360597?v=3"
#> 
#> [[6]]$gravatar_id
#> [1] ""
#> 
#> [[6]]$url
#> [1] "https://api.github.com/users/masalmon"
#> 
#> [[6]]$html_url
#> [1] "https://github.com/masalmon"
#> 
#> [[6]]$followers_url
#> [1] "https://api.github.com/users/masalmon/followers"
#> 
#> [[6]]$following_url
#> [1] "https://api.github.com/users/masalmon/following{/other_user}"
#> 
#> [[6]]$gists_url
#> [1] "https://api.github.com/users/masalmon/gists{/gist_id}"
#> 
#> [[6]]$starred_url
#> [1] "https://api.github.com/users/masalmon/starred{/owner}{/repo}"
#> 
#> [[6]]$subscriptions_url
#> [1] "https://api.github.com/users/masalmon/subscriptions"
#> 
#> [[6]]$organizations_url
#> [1] "https://api.github.com/users/masalmon/orgs"
#> 
#> [[6]]$repos_url
#> [1] "https://api.github.com/users/masalmon/repos"
#> 
#> [[6]]$events_url
#> [1] "https://api.github.com/users/masalmon/events{/privacy}"
#> 
#> [[6]]$received_events_url
#> [1] "https://api.github.com/users/masalmon/received_events"
#> 
#> [[6]]$type
#> [1] "User"
#> 
#> [[6]]$site_admin
#> [1] FALSE
#> 
#> [[6]]$name
#> [1] "Maëlle Salmon"
#> 
#> [[6]]$company
#> [1] "ISGlobal"
#> 
#> [[6]]$blog
#> [1] "https://masalmon.github.io/"
#> 
#> [[6]]$location
#> [1] "Barcelona, Spain"
#> 
#> [[6]]$email
#> NULL
#> 
#> [[6]]$hireable
#> NULL
#> 
#> [[6]]$bio
#> [1] "I use R for getting, preparing, visualizing and analyzing data mostly about air quality & public health."
#> 
#> [[6]]$public_repos
#> [1] 31
#> 
#> [[6]]$public_gists
#> [1] 0
#> 
#> [[6]]$followers
#> [1] 34
#> 
#> [[6]]$following
#> [1] 38
#> 
#> [[6]]$created_at
#> [1] "2014-08-05T08:10:04Z"
#> 
#> [[6]]$updated_at
#> [1] "2016-10-24T09:54:57Z"
#> 
#> 
str(lapply(gh_users, `[`, c("login", "name")))
#> List of 6
#>  $ :List of 2
#>   ..$ login: chr "gaborcsardi"
#>   ..$ name : chr "Gábor Csárdi"
#>  $ :List of 2
#>   ..$ login: chr "jennybc"
#>   ..$ name : chr "Jennifer (Jenny) Bryan"
#>  $ :List of 2
#>   ..$ login: chr "jtleek"
#>   ..$ name : chr "Jeff L."
#>  $ :List of 2
#>   ..$ login: chr "juliasilge"
#>   ..$ name : chr "Julia Silge"
#>  $ :List of 2
#>   ..$ login: chr "leeper"
#>   ..$ name : chr "Thomas J. Leeper"
#>  $ :List of 2
#>   ..$ login: chr "masalmon"
#>   ..$ name : chr "Maëlle Salmon"