Game of Thrones POV characters

Info on the point-of-view (POV) characters from the first five books in the Song of Ice and Fire series by George R. R. Martin. Retrieved from An API Of Ice And Fire.

Usage

got_chars

Format

A unnamed list with 30 components, each representing a POV character. Each character's component is a named list of length 18, containing information such as name, aliases, and house allegiances.

Source

https://anapioficeandfire.com

See also

Examples

str(got_chars, max.level = 1, list.len = 10)
#> List of 30
#>  $ :List of 18
#>  $ :List of 18
#>  $ :List of 18
#>  $ :List of 18
#>  $ :List of 18
#>  $ :List of 18
#>  $ :List of 18
#>  $ :List of 18
#>  $ :List of 18
#>  $ :List of 18
#>   [list output truncated]
str(got_chars[[1]])
#> List of 18
#>  $ url        : chr "https://www.anapioficeandfire.com/api/characters/1022"
#>  $ id         : int 1022
#>  $ name       : chr "Theon Greyjoy"
#>  $ gender     : chr "Male"
#>  $ culture    : chr "Ironborn"
#>  $ born       : chr "In 278 AC or 279 AC, at Pyke"
#>  $ died       : chr ""
#>  $ alive      : logi TRUE
#>  $ titles     : chr [1:2] "Prince of Winterfell" "Lord of the Iron Islands (by law of the green lands)"
#>  $ aliases    : chr [1:4] "Prince of Fools" "Theon Turncloak" "Reek" "Theon Kinslayer"
#>  $ father     : chr ""
#>  $ mother     : chr ""
#>  $ spouse     : chr ""
#>  $ allegiances: chr "House Greyjoy of Pyke"
#>  $ books      : chr [1:3] "A Game of Thrones" "A Storm of Swords" "A Feast for Crows"
#>  $ povBooks   : chr [1:2] "A Clash of Kings" "A Dance with Dragons"
#>  $ tvSeries   : chr [1:6] "Season 1" "Season 2" "Season 3" "Season 4" ...
#>  $ playedBy   : chr "Alfie Allen"
str(lapply(got_chars, `[`, c("name", "culture")))
#> List of 30
#>  $ :List of 2
#>   ..$ name   : chr "Theon Greyjoy"
#>   ..$ culture: chr "Ironborn"
#>  $ :List of 2
#>   ..$ name   : chr "Tyrion Lannister"
#>   ..$ culture: chr ""
#>  $ :List of 2
#>   ..$ name   : chr "Victarion Greyjoy"
#>   ..$ culture: chr "Ironborn"
#>  $ :List of 2
#>   ..$ name   : chr "Will"
#>   ..$ culture: chr ""
#>  $ :List of 2
#>   ..$ name   : chr "Areo Hotah"
#>   ..$ culture: chr "Norvoshi"
#>  $ :List of 2
#>   ..$ name   : chr "Chett"
#>   ..$ culture: chr ""
#>  $ :List of 2
#>   ..$ name   : chr "Cressen"
#>   ..$ culture: chr ""
#>  $ :List of 2
#>   ..$ name   : chr "Arianne Martell"
#>   ..$ culture: chr "Dornish"
#>  $ :List of 2
#>   ..$ name   : chr "Daenerys Targaryen"
#>   ..$ culture: chr "Valyrian"
#>  $ :List of 2
#>   ..$ name   : chr "Davos Seaworth"
#>   ..$ culture: chr "Westeros"
#>  $ :List of 2
#>   ..$ name   : chr "Arya Stark"
#>   ..$ culture: chr "Northmen"
#>  $ :List of 2
#>   ..$ name   : chr "Arys Oakheart"
#>   ..$ culture: chr "Reach"
#>  $ :List of 2
#>   ..$ name   : chr "Asha Greyjoy"
#>   ..$ culture: chr "Ironborn"
#>  $ :List of 2
#>   ..$ name   : chr "Barristan Selmy"
#>   ..$ culture: chr "Westeros"
#>  $ :List of 2
#>   ..$ name   : chr "Varamyr"
#>   ..$ culture: chr "Free Folk"
#>  $ :List of 2
#>   ..$ name   : chr "Brandon Stark"
#>   ..$ culture: chr "Northmen"
#>  $ :List of 2
#>   ..$ name   : chr "Brienne of Tarth"
#>   ..$ culture: chr ""
#>  $ :List of 2
#>   ..$ name   : chr "Catelyn Stark"
#>   ..$ culture: chr "Rivermen"
#>  $ :List of 2
#>   ..$ name   : chr "Cersei Lannister"
#>   ..$ culture: chr "Westerman"
#>  $ :List of 2
#>   ..$ name   : chr "Eddard Stark"
#>   ..$ culture: chr "Northmen"
#>  $ :List of 2
#>   ..$ name   : chr "Jaime Lannister"
#>   ..$ culture: chr "Westerlands"
#>  $ :List of 2
#>   ..$ name   : chr "Jon Connington"
#>   ..$ culture: chr "Stormlands"
#>  $ :List of 2
#>   ..$ name   : chr "Jon Snow"
#>   ..$ culture: chr "Northmen"
#>  $ :List of 2
#>   ..$ name   : chr "Aeron Greyjoy"
#>   ..$ culture: chr "Ironborn"
#>  $ :List of 2
#>   ..$ name   : chr "Kevan Lannister"
#>   ..$ culture: chr ""
#>  $ :List of 2
#>   ..$ name   : chr "Melisandre"
#>   ..$ culture: chr "Asshai"
#>  $ :List of 2
#>   ..$ name   : chr "Merrett Frey"
#>   ..$ culture: chr "Rivermen"
#>  $ :List of 2
#>   ..$ name   : chr "Quentyn Martell"
#>   ..$ culture: chr "Dornish"
#>  $ :List of 2
#>   ..$ name   : chr "Samwell Tarly"
#>   ..$ culture: chr "Andal"
#>  $ :List of 2
#>   ..$ name   : chr "Sansa Stark"
#>   ..$ culture: chr "Northmen"