Entities from the Star Wars Universe

Data retrieved from the swapi API on the Star Wars Universe.

Usage

sw_people

sw_films

sw_planets

sw_species

sw_vehicles

sw_starships

Format

Unnamed lists with varying number of components.

Source

Data originally obtained from http://swapi.co/ using the rwars package: https://github.com/Ironholds/rwars. The Star Wars API appears to have moved to https://pipedream.com/apps/swapi since that time.

Details

sw_people List of individual people or characters within the Star Wars universe.

sw_starships List of transport crafts with hyperdrive capability.

sw_vehicles List of transport crafts without hyperdrive capability.

sw_films List of Star Wars films.

sw_species List of types of people or characters within the Star Wars Universe.

sw_planets List of large masses, planets or planetoids in the Star Wars Universe, at the time of 0 ABY.

Examples

# sw_people
str(sw_people, max.level = 1)
#> List of 87
#>  $ :List of 16
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 15
#>  $ :List of 15
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 15
#>  $ :List of 16
#>  $ :List of 16
#>  $ :List of 14
#>  $ :List of 16
#>  $ :List of 15
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 16
#>  $ :List of 15
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 15
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 15
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 15
#>  $ :List of 14
#>  $ :List of 15
#>  $ :List of 15
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 13
#>  $ :List of 14
#>  $ :List of 16
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 15
#>  $ :List of 14
#>  $ :List of 15
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 15
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 15
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 13
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 16
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 13
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 15
#>  $ :List of 14
#>  $ :List of 13
#>  $ :List of 15
str(sw_people[[1]])
#> List of 16
#>  $ name      : chr "Luke Skywalker"
#>  $ height    : chr "172"
#>  $ mass      : chr "77"
#>  $ hair_color: chr "blond"
#>  $ skin_color: chr "fair"
#>  $ eye_color : chr "blue"
#>  $ birth_year: chr "19BBY"
#>  $ gender    : chr "male"
#>  $ homeworld : chr "http://swapi.co/api/planets/1/"
#>  $ films     : chr [1:5] "http://swapi.co/api/films/6/" "http://swapi.co/api/films/3/" "http://swapi.co/api/films/2/" "http://swapi.co/api/films/1/" ...
#>  $ species   : chr "http://swapi.co/api/species/1/"
#>  $ vehicles  : chr [1:2] "http://swapi.co/api/vehicles/14/" "http://swapi.co/api/vehicles/30/"
#>  $ starships : chr [1:2] "http://swapi.co/api/starships/12/" "http://swapi.co/api/starships/22/"
#>  $ created   : chr "2014-12-09T13:50:51.644000Z"
#>  $ edited    : chr "2014-12-20T21:17:56.891000Z"
#>  $ url       : chr "http://swapi.co/api/people/1/"
sapply(sw_people, `[[`, "name")
#>  [1] "Luke Skywalker"        "C-3PO"                 "R2-D2"                
#>  [4] "Darth Vader"           "Leia Organa"           "Owen Lars"            
#>  [7] "Beru Whitesun lars"    "R5-D4"                 "Biggs Darklighter"    
#> [10] "Obi-Wan Kenobi"        "Anakin Skywalker"      "Wilhuff Tarkin"       
#> [13] "Chewbacca"             "Han Solo"              "Greedo"               
#> [16] "Jabba Desilijic Tiure" "Wedge Antilles"        "Jek Tono Porkins"     
#> [19] "Yoda"                  "Palpatine"             "Boba Fett"            
#> [22] "IG-88"                 "Bossk"                 "Lando Calrissian"     
#> [25] "Lobot"                 "Ackbar"                "Mon Mothma"           
#> [28] "Arvel Crynyd"          "Wicket Systri Warrick" "Nien Nunb"            
#> [31] "Qui-Gon Jinn"          "Nute Gunray"           "Finis Valorum"        
#> [34] "Jar Jar Binks"         "Roos Tarpals"          "Rugor Nass"           
#> [37] "Ric Olié"              "Watto"                 "Sebulba"              
#> [40] "Quarsh Panaka"         "Shmi Skywalker"        "Darth Maul"           
#> [43] "Bib Fortuna"           "Ayla Secura"           "Dud Bolt"             
#> [46] "Gasgano"               "Ben Quadinaros"        "Mace Windu"           
#> [49] "Ki-Adi-Mundi"          "Kit Fisto"             "Eeth Koth"            
#> [52] "Adi Gallia"            "Saesee Tiin"           "Yarael Poof"          
#> [55] "Plo Koon"              "Mas Amedda"            "Gregar Typho"         
#> [58] "Cordé"                 "Cliegg Lars"           "Poggle the Lesser"    
#> [61] "Luminara Unduli"       "Barriss Offee"         "Dormé"                
#> [64] "Dooku"                 "Bail Prestor Organa"   "Jango Fett"           
#> [67] "Zam Wesell"            "Dexter Jettster"       "Lama Su"              
#> [70] "Taun We"               "Jocasta Nu"            "Ratts Tyerell"        
#> [73] "R4-P17"                "Wat Tambor"            "San Hill"             
#> [76] "Shaak Ti"              "Grievous"              "Tarfful"              
#> [79] "Raymus Antilles"       "Sly Moore"             "Tion Medon"           
#> [82] "Finn"                  "Rey"                   "Poe Dameron"          
#> [85] "BB8"                   "Captain Phasma"        "Padmé Amidala"        

# sw_films
str(sw_films, max.level = 1)
#> List of 7
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 14
#>  $ :List of 13
str(sw_films[[1]])
#> List of 14
#>  $ title        : chr "A New Hope"
#>  $ episode_id   : int 4
#>  $ opening_crawl: chr "It is a period of civil war.\r\nRebel spaceships, striking\r\nfrom a hidden base, have won\r\ntheir first victo"| __truncated__
#>  $ director     : chr "George Lucas"
#>  $ producer     : chr "Gary Kurtz, Rick McCallum"
#>  $ release_date : chr "1977-05-25"
#>  $ characters   : chr [1:18] "http://swapi.co/api/people/1/" "http://swapi.co/api/people/2/" "http://swapi.co/api/people/3/" "http://swapi.co/api/people/4/" ...
#>  $ planets      : chr [1:3] "http://swapi.co/api/planets/2/" "http://swapi.co/api/planets/3/" "http://swapi.co/api/planets/1/"
#>  $ starships    : chr [1:8] "http://swapi.co/api/starships/2/" "http://swapi.co/api/starships/3/" "http://swapi.co/api/starships/5/" "http://swapi.co/api/starships/9/" ...
#>  $ vehicles     : chr [1:4] "http://swapi.co/api/vehicles/4/" "http://swapi.co/api/vehicles/6/" "http://swapi.co/api/vehicles/7/" "http://swapi.co/api/vehicles/8/"
#>  $ species      : chr [1:5] "http://swapi.co/api/species/5/" "http://swapi.co/api/species/3/" "http://swapi.co/api/species/2/" "http://swapi.co/api/species/1/" ...
#>  $ created      : chr "2014-12-10T14:23:31.880000Z"
#>  $ edited       : chr "2015-04-11T09:46:52.774897Z"
#>  $ url          : chr "http://swapi.co/api/films/1/"
sapply(sw_films, `[[`, "title")
#> [1] "A New Hope"              "Attack of the Clones"   
#> [3] "The Phantom Menace"      "Revenge of the Sith"    
#> [5] "Return of the Jedi"      "The Empire Strikes Back"
#> [7] "The Force Awakens"